+++ index.html
@@ -8,8 +8,8 @@
 		/>
 		<title>News stand</title>
 		<link rel="stylesheet" href="css/style.css" />
-		<script defer type="module" src="js/readJSON.js"></script>
-		<script defer type="module" src="js/grid.js"></script>
+		<script defer type="module" src="js/utils/readJSON.js"></script>
+		<script defer type="module" src="js/press.js"></script>
 		<!-- <script defer type="module" src="js/test.js"></script> -->
 		<script
 			defer
